Add AggregateDao and add accessor to PointValueDao

+++ Core/src/com/serotonin/m2m2/rt/script/AnalogStatisticsWrapper.java
@@ -77,7 +77,7 @@ public class AnalogStatisticsWrapper {
     }
 
     public int getCount() {
-        return statistics.getCount();
+        return (int) statistics.getCount();
     }
 
     public double getDelta() {
